Alpha Version of Login Page
I added a login page that includes a title, email form, password form, and submit form. There is also a ucf logo and a link to UCF's website. If you enter a valid email and password the page redirects to another html page that will serve as the home page.