Added the new versions of Home and Login .css
The Home.css implements some of the design choices we draw up on our storyboard. It has additional colors and padding for different sections of the menus. 

The Login.css uses more of the Bootstrap design methods for login's and padding. This also supports the use of assets like etc, a company logo or in this case a school logo.