Update home.html
This update adds in our proposed idea's to the site. This adds a functioning sidebar with icons and words that follow up after clicking on the hamburger menu. An interactive calendar with the assignment view to go along with it, as well as a color theme (subject to change tbh)